requirements.txt
streamlit==1.45.1
gspread==6.2.1
pandas==2.2.3
python-dotenv==1.0.0
matplotlib==3.10.3
streamlit_autorefresh==1.0.1